feat(card): add selection card component
