+++ core/index.type.tsx
@@ -82,5 +82,4 @@ export { ChoiceListProps } from './components/organisms/choiceList';
 export { DividerProps } from './components/atoms/divider';
 export { HelpTextProps } from './components/atoms/helpText';
 export { LinkButtonProps } from './components/atoms/linkButton';
-export { CardItemProps } from './components/atoms/selectionCard/CardItem';
 export { SelectionCardProps } from './components/atoms/selectionCard';
